Trying to fix the 2D vector dereference, no luck yet, pick it up on the other side for more work.

+++ SudokuSolver/SudokuSolver/SudokuCell.h
@@ -74,7 +74,7 @@ int SudokuCell::getVal() {
 }
 
 int SudokuCell::getSize() {
-	return notes.size();
+	return (int) notes.size();
 }
 
 void SudokuCell::setNote(int index) {
